server/requirements.txt
Flask==2.0.2
Flask_Login==0.5.0
Flask_RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
requests==2.26.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
flask-security-too 
redis
celery[redis]
Flask-Caching